move cert signing tests
